requirements.txt
numpy
pyaudio
rich
textual
google-generativeai
pillow
requests
pyreadline3 
pywin32